requirements.txt
llama-index==0.10.27
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.27
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.14
llama-index-llms-openai-like==0.1.3
llama-index-llms-vllm==0.1.7
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-faiss==0.1.2
llama-index-vector-stores-milvus==0.1.10
llama-parse==0.4.0
llamaindex-py-client==0.1.16
docx2txt==0.8